Fix FAMD slicing issue

+++ prince/ca.py
@@ -97,7 +97,7 @@ class CA(base.BaseEstimator, base.TransformerMixin):
         _, row_names, _, _ = util.make_labels_and_names(X)
 
         if isinstance(X, pd.SparseDataFrame):
-            X = X.to_coo()
+            X = X.to_coo().astype(float)
         elif isinstance(X, pd.DataFrame):
             X = X.to_numpy()
